SOW-2. Milestone 3 (final)

+++ examples/coap/libcoap_build.sh
@@ -4,29 +4,34 @@
 set -e
 
 LOG_FILE="build.log"
+PROJECT_DIR="libcoap"
 BUILD_DIR="build"
 
-cd libcoap
+cd "$PROJECT_DIR"
 
-# Remove the existing directory if it exists
-if [ -d "$BUILD_DIR" ]; then
-    rm -rf "$BUILD_DIR"
+# if libcoap is not already built, then build else return the path to the static library
+if [ ! -f "$BUILD_DIR/include/coap3/coap.h" ]; then
+    # Remove the existing directory if it exists
+    if [ -d "$BUILD_DIR" ]; then
+        rm -rf "$BUILD_DIR"
+    fi
+
+    # Create a build directory
+    mkdir "$BUILD_DIR"
+    cd "$BUILD_DIR"
+
+    # Configure the build for a static library and append output to log file
+    cmake .. -DCMAKE_BUILD_TYPE=MinSizeRel -DBUILD_SHARED_LIBS=ON -DENABLE_DTLS=ON -DWITH_OPENSSL=ON -DENABLE_SERVER_MODE=OFF -DENABLE_EXAMPLES=OFF -DENABLE_DOCS=OFF >> "$LOG_FILE" 2>&1
+
+    # Build the library and append output to the same log file
+    make >> "$LOG_FILE" 2>&1
+
+    # Copy the include directory to the build directory
+    cp -r ../include ./
+
+    # Return to the parent directory
+    cd ..
 fi
 
-# Create a build directory
-mkdir "$BUILD_DIR"
-cd "$BUILD_DIR"
-
-# Configure the build for a static library and append output to log file
-cmake .. -DCMAKE_BUILD_TYPE=MinSizeRel -DBUILD_SHARED_LIBS=ON -DENABLE_DTLS=ON -DWITH_TINYDTLS=ON -DENABLE_SERVER_MODE=OFF -DENABLE_EXAMPLES=OFF -DENABLE_DOCS=OFF >> "$LOG_FILE" 2>&1
-
-# Build the library and append output to the same log file
-make >> "$LOG_FILE" 2>&1
-
-# Copy the include directory to the build directory
-cp -r ../include ./
-
-# The static library will be located in build/src/
-LIB_PATH="$(pwd)"
 # Echo the path to the static library
-echo -n "$LIB_PATH"
+echo -n "$(pwd)/$BUILD_DIR"
